dailybuild: EmbPython ライブラリ: ソース登録漏れ対応
git-svn-id: http://springhead.info/spr2/Springhead2/trunk@6879 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Samples/EmbPython/MyEPApp/MyEPApp.h
@@ -56,8 +56,8 @@ public:
 
 	virtual void OnStep(){
 		UTAutoLock critical(EPCriticalSection);
-		spaceNavigator0->Update(GetFWScene()->GetPHScene()->GetTimeStep());
-		spaceNavigator1->Update(GetFWScene()->GetPHScene()->GetTimeStep());
+		spaceNavigator0->Update((float) GetFWScene()->GetPHScene()->GetTimeStep());
+		spaceNavigator1->Update((float) GetFWScene()->GetPHScene()->GetTimeStep());
 		GetFWScene()->Step();
 	}
